-- Huge changes all over the place to get the full data set reading in about 16 seconds on a 7200RPM hard drive; I don't think hard drive is actually the bottleneck but there's not a lot more I can do there without taking some ridiculous measures (like rewriting the DB I/O from the ground up) -- One of the casualties of my push for efficiency was the Date field in UserRating; surprisingly this didn't end up improving the read speed by a huge amount (probably on the order of 2-3 seconds) but it does lower the profile of the data on disk by about 50% so I'm leaving it as-is for now

+++ knn/NaiveKNN.java
@@ -54,7 +54,7 @@ public class NaiveKNN implements Predictor {
 			}
 			
 			// user ratings from movie A
-			HashMap<Integer, Integer> aRatings = new HashMap<Integer, Integer>();
+			HashMap<Integer, Byte> aRatings = new HashMap<Integer, Byte>();
 			
 			for(UserRating r: ratings) {
 				aRatings.put(r.userId, r.rating);
@@ -69,7 +69,7 @@ public class NaiveKNN implements Predictor {
 				
 				// System.out.println("otherRatings.size(): " + otherRatings.size());
 				
-				HashMap<Integer, Integer> bRatings = new HashMap<Integer, Integer>();
+				HashMap<Integer, Byte> bRatings = new HashMap<Integer, Byte>();
 				
 				// now comparing one movie's set of ratings with another's
 				for(UserRating r : otherRatings) {
@@ -79,7 +79,7 @@ public class NaiveKNN implements Predictor {
 				int ratingTotal = 0;
 				int ratingCounter = 0;
 				
-				for(Map.Entry<Integer, Integer> entry : aRatings.entrySet()){
+				for(Map.Entry<Integer, Byte> entry : aRatings.entrySet()){
 					if(bRatings.containsKey(entry.getKey()) ) {
 						int squareDiff = entry.getValue() - bRatings.get(entry.getKey());
 						squareDiff = squareDiff * squareDiff;
